# backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Social Media API",
    "description": "API for app </b>Social Media</b>"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Please enter a valid token to test the requests below..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/users/": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/users/{id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/users/friends/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/users/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/users/{id}/follow": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/users/{id}/unfollow": {
      "put": {
        "tags": [
          "Users"
        ],
        "description": "Endpoint for users.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/admin/users/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint for messages.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "textSearch",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Data invalid "
          },
          "404": {
            "description": "Not found "
          },
          "500": {
            "description": "Error Internal Server"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
...
}